requirements.txt
flask==3.0.2
requests==2.31.0
google-cloud-storage==2.16.0
gunicorn==21.2.0